Added better loading indicators. Limited postpone to 4 weeks. Added unskip recurrence

+++ src/components/modals/SkipRecurrence.js
@@ -1,12 +1,13 @@
 import { ModalContent, ModalView } from '../Modal'
 import { ButtonStyled } from '@open-tender/components'
 import { closeModal, toggleSidebar } from '../../slices'
-import React from 'react'
+import React, { useEffect } from 'react'
 import { useDispatch, useSelector } from 'react-redux'
 import {
-  editOrder,
+  addItemToCart, deleteCustomerOrder,
+  editOrder, incrementItemInCart,
   removeItemFromCartById, resetCheckout,
-  resetOrder,
+  resetOrder, selectCart, selectOrder,
   selectTimezone,
   setSubmitting, showNotification,
   submitOrder
@@ -19,14 +20,30 @@ const SkipRecurrence = ({ recurrence, order }) => {
   const readableDate = isoToDateStr(order.requested_at, tz, 'EEEE, MMMM dd')
   const dispatch = useDispatch()
 
-  const skipRecurrence = async () => {
+  const toggleSkipRecurrence = async () => {
     await dispatch(editOrder(order, false))
 
-    dispatch(removeItemFromCartById(recurrence.item))
-    dispatch(setSubmitting(true))
-    dispatch(submitOrder())
+    if (recurrence.isSkipped) {
+      dispatch(addItemToCart(
+        {...recurrence.item,
+          groups: [],
+          quantity: recurrence.quantity,
+          frequency: recurrence.frequency,
+          recurrence_id: recurrence.id
+        }))
 
-    dispatch(showNotification(`${recurrence.item.name} will be skipped on your upcoming order on ${readableDate}`))
+    } else {
+      const cart = dispatch(removeItemFromCartById(recurrence.item))
+
+      if (cart.length === 0) {
+        dispatch(deleteCustomerOrder(order.order_id))
+      }
+    }
+
+    dispatch(setSubmitting(true))
+    await dispatch(submitOrder())
+
+    dispatch(showNotification(`${recurrence.item.name} was ${recurrence.isSkipped?'re-added to':'skipped on'} your upcoming order on ${readableDate}`))
     dispatch(resetOrder())
     dispatch(resetCheckout())
   }
@@ -34,31 +51,31 @@ const SkipRecurrence = ({ recurrence, order }) => {
   return (
     <ModalView>
       <ModalContent
-        title='Are you sure you want to skip this item in the next order?'
+        title={`Are you sure you want to ${recurrence.isSkipped? 'un-skip':'skip'} this item in the next order?`}
         subtitle={
           <>
             <p>
-              {recurrence.item.name} will be removed from the upcoming order on {readableDate}.
+              {recurrence.item.name} will be {recurrence.isSkipped? 're-added to':'removed from'} the upcoming order on {readableDate}.
             </p>
-            <p>
-              It will continue to be added to future recurring orders.
-            </p>
+            { !recurrence.isSkipped &&
+              <p>
+                It will continue to be added to future recurring orders.
+              </p>
+            }
           </>
         }
         footer={
           <div>
-            <ButtonStyled onClick={skipRecurrence}>
-              Skip Next Order
+            <ButtonStyled onClick={toggleSkipRecurrence}>
+              {recurrence.isSkipped && 'Un-'}Skip This Item
             </ButtonStyled>
             <ButtonStyled onClick={() => dispatch(closeModal())} color='secondary'>
               Cancel
             </ButtonStyled>
           </div>
         }
-
       >
       </ModalContent>
-      }
     </ModalView>
   )
 }
